Shader descriptor heaps are now resized if too small and pooled to prevent the nvidia driver from getting stuck sometimes...

+++ Modules/D3D12/Sources/RND3D12Framebuffer.cpp
@@ -378,7 +378,7 @@ namespace RN
 		if(_colorTargets.size() == 0)
 			return;
 
-		D3D12_RECT clearRect{ 0, 0, GetSize().x, GetSize().y };
+		D3D12_RECT clearRect{ 0, 0, static_cast<LONG>(GetSize().x), static_cast<LONG>(GetSize().y) };
 		commandList->GetCommandList()->ClearRenderTargetView(*_rtvHandle, &color.r, 1, &clearRect);
 	}
